Fixed some minor warnings
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@1804 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/SColor.h
@@ -15,7 +15,7 @@ namespace video
 	//! Creates a 16 bit A1R5G5B5 color
 	inline u16 RGBA16(u32 r, u32 g, u32 b, u32 a=0xFF)
 	{
-		return ((a & 0x80) << 8 |
+		return (u16)((a & 0x80) << 8 |
 			(r & 0xF8) << 7 |
 			(g & 0xF8) << 2 |
 			(b & 0xF8) >> 3);
@@ -42,7 +42,7 @@ namespace video
 	//! Converts a 32bit (X8R8G8B8) color to a 16bit A1R5G5B5 color
 	inline u16 X8R8G8B8toA1R5G5B5(u32 color)
 	{
-		return (0x8000 |
+		return (u16)(0x8000 |
 			( color & 0x00F80000) >> 9 |
 			( color & 0x0000F800) >> 6 |
 			( color & 0x000000F8) >> 3);
@@ -52,7 +52,7 @@ namespace video
 	//! Converts a 32bit (A8R8G8B8) color to a 16bit A1R5G5B5 color
 	inline u16 A8R8G8B8toA1R5G5B5(u32 color)
 	{
-		return (( color & 0x80000000) >> 16|
+		return (u16)(( color & 0x80000000) >> 16|
 			( color & 0x00F80000) >> 9 |
 			( color & 0x0000F800) >> 6 |
 			( color & 0x000000F8) >> 3);
@@ -62,7 +62,7 @@ namespace video
 	//! Converts a 32bit (A8R8G8B8) color to a 16bit R5G6B5 color
 	inline u16 A8R8G8B8toR5G6B5(u32 color)
 	{
-		return (( color & 0x00F80000) >> 8 |
+		return (u16)(( color & 0x00F80000) >> 8 |
 			( color & 0x0000FC00) >> 5 |
 			( color & 0x000000F8) >> 3);
 	}
@@ -232,10 +232,10 @@ namespace video
 		\param dest: address where the 4x8 bit OpenGL color is stored. */
 		void toOpenGLColor(u8* dest) const
 		{
-			*dest =   getRed();
-			*++dest = getGreen();
-			*++dest = getBlue();
-			*++dest = getAlpha();
+			*dest =   (u8)getRed();
+			*++dest = (u8)getGreen();
+			*++dest = (u8)getBlue();
+			*++dest = (u8)getAlpha();
 		}
 
 		//! Sets all four components of the color at once.
